+++ src/main/java/module-info.java
@@ -1,0 +1,7 @@
+module silly.strings {
+	requires org.apache.commons.lang3;
+	requires silly.collection.main;
+	exports io.github.astrapi69.regex;
+	exports io.github.astrapi69.string;
+
+}